+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/ArrayList.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5cd38a744bd07ff4bb34909e0c68a93f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
