+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/ArrayList/ArrayList.Enumerators.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cc918666891cc0143a50182dbca52605
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
