+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/ArrayList/ArrayList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 97a9cba1f4430c44085be42b7a3df632
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
