+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/Enumerators.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2ac78f1908396f143838c99559e952e3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
