+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/Enumerators/IInFunc.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0bffa7b5b14061c4dade50e0fe0c0a7a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
