+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/Enumerators/IMutEnumerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a82569689000dd743976350ddfc3fb8d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
