+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/Enumerators/WhereMutEnumerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 32d917cd85e2d264ab3a512aac7b52dd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
