+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/Memory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a8bd0fb3f687bd7419e2aed4da3c2234
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
