+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/Memory/AllocatorUtilities.cs
@@ -1,0 +1,10 @@
+namespace UnityEditor.Rendering.HighDefinition
+{
+    public unsafe static class AllocatorUtilities
+    {
+        public static T* Allocate<T, A>(this A allocator)
+            where T: unmanaged
+            where A : IMemoryAllocator
+            => (T*)allocator.Allocate((ulong)sizeof(T));
+    }
+}
