+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/Memory/AllocatorUtilities.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be4a8ba16a62c6c4891eca61e08cfd58
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
