+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/Memory/FixedAllocator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fbfb475bc8762c848aebe9959d75fa79
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
