+++ com.unity.render-pipelines.high-definition/Runtime/Utilities/Memory/IMemoryAllocator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e56fd21284d416847a7c12d1e2a4a273
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
